feat: add sample charts
